requirements.txt
flask
berserk
python-chess
stockfish
numpy
torch